parallelize ast parsing and puppeteer browser page setup
both (chains of) operation(s) can take several 100 ms (ast parsing only on larger stylesheets),
so by executing them in parallel time can be saved.
