+++ src/index.js
@@ -1,5 +1,4 @@
 import fs from 'fs'
-import csstree from 'css-tree'
 import puppeteer from 'puppeteer'
 import debug from 'debug'
 
@@ -100,33 +99,9 @@ function prepareForceIncludeForSerialization (forceInclude = []) {
   })
 }
 
-const astFromCss = async function astFromCss (options) {
-  // breaks puppeteer
-  const css = options.cssString.replace(/￿/g, '\f042')
-
-  let parsingErrors = []
-  let ast = csstree.parse(css, {
-    onParseError: error => parsingErrors.push(error.formattedMessage)
-  })
-  debuglog(`parsed ast (with ${parsingErrors.length} errors)`)
-
-  if (parsingErrors.length && options.strict === true) {
-    // NOTE: only informing about first error, even if there were more than one.
-    const parsingErrorMessage = parsingErrors[0]
-    throw new Error(
-      `AST parser (css-tree) found ${parsingErrors.length} errors in CSS.
-    Breaking because in strict mode.
-    The first error was:
-    ` + parsingErrorMessage
-    )
-  }
-  return ast
-}
-
 // const so not hoisted, so can get regeneratorRuntime inlined above, needed for Node 4
 const generateCriticalCssWrapped = async function generateCriticalCssWrapped (
   options,
-  ast,
   { forceTryRestartBrowser } = {}
 ) {
   const width = parseInt(options.width || DEFAULT_VIEWPORT_WIDTH, 10)
@@ -168,10 +143,11 @@ const generateCriticalCssWrapped = async function generateCriticalCssWrapped (
       formattedCss = await generateCriticalCss({
         browser,
         url: options.url,
-        ast,
+        cssString: options.cssString,
         width,
         height,
         forceInclude,
+        strict: options.strict,
         userAgent: options.userAgent || DEFAULT_USER_AGENT,
         renderWaitTime: options.renderWaitTime || DEFAULT_RENDER_WAIT_TIMEOUT,
         timeout: timeoutWait,
@@ -207,8 +183,8 @@ const generateCriticalCssWrapped = async function generateCriticalCssWrapped (
             (_browserPagesOpen + 1) +
             '\nurl: ' +
             options.url +
-            '\nAST children: ' +
-            ast.children.getSize()
+            '\ncss length: ' +
+            options.cssString.length
         )
         // for some reason Chromium is no longer opened;
         // perhaps it crashed
@@ -222,7 +198,7 @@ const generateCriticalCssWrapped = async function generateCriticalCssWrapped (
         }
         // retry
         resolve(
-          generateCriticalCssWrapped(options, ast, {
+          generateCriticalCssWrapped(options, {
             forceTryRestartBrowser: true
           })
         )
@@ -296,8 +272,7 @@ module.exports = function (options, callback) {
       getBrowser: options.puppeteer && options.puppeteer.getBrowser
     })
     try {
-      const ast = await astFromCss(options)
-      const criticalCss = await generateCriticalCssWrapped(options, ast)
+      const criticalCss = await generateCriticalCssWrapped(options)
       cleanupAndExit({ returnValue: criticalCss })
     } catch (err) {
       cleanupAndExit({ error: err })
